pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-tts-ru"
version = "0.1.0"
description = "Local Russian PDF to audio pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "OpenAI ChatGPT scaffold" }
]
dependencies = [
  "PyMuPDF>=1.27.0",
  "piper-tts>=1.3.0",
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]
silero = [
  "silero>=0.5.0",
  "torch>=2.0.0",
]

[project.scripts]
pdf-tts-ru = "pdf_tts_ru.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
